Declare Year plugin for Solr search index
We needed to declare a new Solr calculated field as a plugin to a Drupal module.  This archive module is the closest one we have to being relevant.

Ultimately, we'll probably want to move this declaration to a separate COB Search module.  But, right now, we just need to declare this one field type, and do not want to create a whole new Drupal module to do it.

Updates #7